requirements.txt
altgraph==0.17.4
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
dlib==20.0.0
deepface==0.0.79
docutils==0.21.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
geocoder==1.38.1
geopy==2.4.1
huggingface-hub==0.32.4
idna==3.10
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.6
Keras==2.10.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivy_deps.gstreamer==0.3.4
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
openrouteservice==2.3.3
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==10.4.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
pypiwin32==223
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
SpeechRecognition==3.14.3
srt==3.5.3
sympy==1.14.0
tensorflow==2.10.0
timm==1.0.15
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
vosk==0.3.45

